Began refactoring MLPOptimizer
Instead of having multiple optimizer, we now only use a single class MLPOptimizer which can take a batch size as argument.

On top of that, batch scheduling has now its own class BatchScheduler, which allow the user to control how batch are treated.
This also separate the complex logic of creating thread and splitting the batches in a separate file.

Added the MLPWeightUpdater class. This class is responsible for summing the gradient, and applying the optimization function at the end of the batch.